requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
langchain==0.1.20
langchain-community==0.0.38
ollama==0.1.7
pypdf2==3.0.1
python-docx==0.8.11
streamlit==1.32.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
json5==0.9.14
langdetect==1.0.9